Added new test class, for calculated values and changeable variables, moved toString from BasicMonster to Monster
